Fixup sorted teams
