+++ shared/profile/showcase-team-offer/container.js
@@ -5,19 +5,19 @@ import {branch, compose, connect, lifecycle, type TypedState} from '../../util/c
 import * as TeamsGen from '../../actions/teams-gen'
 import {HeaderHoc} from '../../common-adapters'
 import {isMobile} from '../../constants/platform'
-import {sortTeamnames} from '../../util/teams'
+import {getSortedTeamnames} from '../../constants/teams'
 
 const mapStateToProps = (state: TypedState) => {
   return {
     _teamNameToIsOpen: state.teams.getIn(['teamNameToIsOpen'], I.Map()),
     _teammembercounts: state.teams.getIn(['teammembercounts'], I.Map()),
-    _teamnames: state.teams.getIn(['teamnames'], I.Set()),
     _teamNameToCanPerform: state.teams.getIn(['teamNameToCanPerform'], I.Map()),
     _teamNameToPublicitySettings: state.teams.getIn(['teamNameToPublicitySettings'], I.Map()),
     _teamNameToAllowPromote: state.teams.getIn(['teamNameToAllowPromote'], I.Map()),
     _teamNameToIsShowcasing: state.teams.getIn(['teamNameToIsShowcasing'], I.Map()),
     _teamNameToRole: state.teams.getIn(['teamNameToRole'], I.Map()),
     _waiting: state.waiting,
+    teamnames: getSortedTeamnames(state),
   }
 }
 
@@ -28,9 +28,6 @@ const mapDispatchToProps = (dispatch: Dispatch, {navigateUp}) => ({
 })
 
 const mergeProps = (stateProps, dispatchProps) => {
-  let teamnames = stateProps._teamnames.toArray()
-  teamnames.sort(sortTeamnames)
-
   return {
     ...stateProps,
     ...dispatchProps,
@@ -39,7 +36,6 @@ const mergeProps = (stateProps, dispatchProps) => {
     teamNameToAllowPromote: stateProps._teamNameToAllowPromote.toObject(),
     teamNameToIsShowcasing: stateProps._teamNameToIsShowcasing.toObject(),
     teamNameToRole: stateProps._teamNameToRole.toObject(),
-    teamnames,
     title: 'Publish your teams',
     waiting: stateProps._waiting.toObject(),
   }
