+++ shared/teams/container.js
@@ -7,7 +7,7 @@ import Teams from './main'
 import openURL from '../util/open-url'
 import {navigateAppend} from '../actions/route-tree'
 import {compose, lifecycle, type TypedState, connect} from '../util/container'
-import {sortTeamnames} from '../util/teams'
+import {getSortedTeamnames} from '../constants/teams'
 import {type Teamname} from '../constants/types/teams'
 
 const mapStateToProps = (state: TypedState) => ({
@@ -15,10 +15,10 @@ const mapStateToProps = (state: TypedState) => ({
   _newTeams: state.teams.getIn(['newTeams'], I.Set()),
   _teamNameToIsOpen: state.teams.getIn(['teamNameToIsOpen'], I.Map()),
   _teammembercounts: state.teams.getIn(['teammembercounts'], I.Map()),
-  _teamnames: state.teams.getIn(['teamnames'], I.Set()),
   _teamresetusers: state.teams.getIn(['teamNameToResetUsers'], I.Map()),
   loaded: state.teams.getIn(['loaded'], false),
   sawChatBanner: state.teams.getIn(['sawChatBanner'], false),
+  teamnames: getSortedTeamnames(state),
 })
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
@@ -47,9 +47,6 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
 })
 
 const mergeProps = (stateProps, dispatchProps) => {
-  let teamnames = stateProps._teamnames.toArray()
-  teamnames.sort(sortTeamnames)
-
   return {
     loaded: stateProps.loaded,
     newTeamRequests: stateProps._newTeamRequests.toArray(),
@@ -57,7 +54,7 @@ const mergeProps = (stateProps, dispatchProps) => {
     sawChatBanner: stateProps.sawChatBanner,
     teamNameToIsOpen: stateProps._teamNameToIsOpen.toObject(),
     teammembercounts: stateProps._teammembercounts.toObject(),
-    teamnames,
+    teamnames: stateProps.teamnames,
     teamresetusers: stateProps._teamresetusers.toObject(),
     ...dispatchProps,
   }
